Update Radius to Area Assignment.py

+++ 1.9.25/Radius to Area Assignment.py
@@ -3,6 +3,7 @@
 #6. Write a python program that accepts radius of a circle and prints its area.
 
 radius = input('Enter a radius: ')
-radius = int(radius)
+radius = float(radius)
 radius_squared = (radius * radius)
 print('The area is:',3.14 * radius_squared)
+
